# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LARAVEL Rest API DOCUMENTATION",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "omarzablah20@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/get_addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get all addresses",
                "description": "",
                "operationId": "getAddresses",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/add_address": {
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Add new address",
                "description": "",
                "operationId": "addAddress",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "region_id",
                                    "city_id",
                                    "township_id"
                                ],
                                "properties": {
                                    "region_id": {
                                        "description": "Enter your region",
                                        "type": "string"
                                    },
                                    "city_id": {
                                        "description": "Enter your city",
                                        "type": "string"
                                    },
                                    "township_id": {
                                        "description": "Enter your township",
                                        "type": "string"
                                    },
                                    "details": {
                                        "description": "Enter your details",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/get_ads": {
            "get": {
                "tags": [
                    "Ads"
                ],
                "summary": "Get Ads",
                "description": "",
                "operationId": "getAds",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login to system",
                "operationId": "login",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register",
                "description": "Register to system",
                "operationId": "register",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "phone",
                                    "password",
                                    "password_confirmation",
                                    "street",
                                    "region_id",
                                    "city_id",
                                    "township_id"
                                ],
                                "properties": {
                                    "firstName": {
                                        "description": "Enter your firstName",
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "description": "Enter your lastName",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Enter your email",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Enter your phone number",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Enter your password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm your password",
                                        "type": "string"
                                    },
                                    "street": {
                                        "description": "Enter your street",
                                        "type": "string"
                                    },
                                    "region_id": {
                                        "description": "Enter your region",
                                        "type": "integer"
                                    },
                                    "city_id": {
                                        "description": "Enter your city",
                                        "type": "integer"
                                    },
                                    "township_id": {
                                        "description": "Enter your township",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Logout from profile",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh",
                "description": "",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/get_catagories": {
            "get": {
                "tags": [
                    "Catagories"
                ],
                "summary": "Get all catagories",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "getCatagories",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/add_order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Add new order",
                "description": "Add a new order for the authenticated user.",
                "operationId": "addOrder",
                "requestBody": {
                    "description": "JSON object containing items to be added to the order",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "description": "Array of items to be added to the order",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "item_id": {
                                                    "description": "ID of the item",
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "description": "Quantity of the item",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The order added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The order added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/get_products/{subCatagoryId}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products by subCatagory",
                "description": "",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "subCatagoryId",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/get_profile_details/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get profile info",
                "description": "",
                "operationId": "getProfileDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/get_regions": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get all regions",
                "description": "",
                "operationId": "getRegions",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/get_cities/{region_id}": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get all cities",
                "description": "",
                "operationId": "getCities",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/get_townships/{city_id}": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get all townships",
                "description": "",
                "operationId": "getTownships",
                "parameters": [
                    {
                        "name": "city_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}